Removed spaces after cast and type values in #define.

+++ AirshipLib/AirshipLogicTests/UAAnalyticsTest.m
@@ -169,57 +169,57 @@
 - (void)testSetMaxTotalDBSize {
     // Set a value higher then the max, should set to the max
     self.analytics.maxTotalDBSize = kMaxTotalDBSizeBytes + 1;
-    STAssertEquals(self.analytics.maxTotalDBSize, (NSInteger) kMaxTotalDBSizeBytes, @"maxTotalDBSize is able to be set above the max value");
+    STAssertEquals(self.analytics.maxTotalDBSize, kMaxTotalDBSizeBytes, @"maxTotalDBSize is able to be set above the max value");
 
     // Set a value lower then then min, should set to the min
     self.analytics.maxTotalDBSize = kMinTotalDBSizeBytes - 1;
-    STAssertEquals(self.analytics.maxTotalDBSize, (NSInteger) kMinTotalDBSizeBytes, @"maxTotalDBSize is able to be set below the min value");
+    STAssertEquals(self.analytics.maxTotalDBSize, kMinTotalDBSizeBytes, @"maxTotalDBSize is able to be set below the min value");
 
     // Set a value between
     self.analytics.maxTotalDBSize = kMinTotalDBSizeBytes + 1;
-    STAssertEquals(self.analytics.maxTotalDBSize, (NSInteger) kMinTotalDBSizeBytes + 1, @"maxTotalDBSize is unable to be set to a valid value");
+    STAssertEquals(self.analytics.maxTotalDBSize, kMinTotalDBSizeBytes + 1, @"maxTotalDBSize is unable to be set to a valid value");
 }
 
 - (void)testSetMaxBatchSize {
     // Set a value higher then the max, should set to the max
     self.analytics.maxBatchSize = kMaxBatchSizeBytes + 1;
-    STAssertEquals(self.analytics.maxBatchSize, (NSInteger) kMaxBatchSizeBytes, @"maxBatchSize is able to be set above the max value");
+    STAssertEquals(self.analytics.maxBatchSize, kMaxBatchSizeBytes, @"maxBatchSize is able to be set above the max value");
 
     // Set a value lower then then min, should set to the min
     self.analytics.maxBatchSize = kMinBatchSizeBytes - 1;
-    STAssertEquals(self.analytics.maxBatchSize, (NSInteger) kMinBatchSizeBytes, @"maxBatchSize is able to be set below the min value");
+    STAssertEquals(self.analytics.maxBatchSize, kMinBatchSizeBytes, @"maxBatchSize is able to be set below the min value");
 
     // Set a value between
     self.analytics.maxBatchSize = kMinBatchSizeBytes + 1;
-    STAssertEquals(self.analytics.maxBatchSize, (NSInteger) kMinBatchSizeBytes + 1, @"maxBatchSize is unable to be set to a valid value");
+    STAssertEquals(self.analytics.maxBatchSize, kMinBatchSizeBytes + 1, @"maxBatchSize is unable to be set to a valid value");
 }
 
 - (void)testSetMaxWait {
     // Set a value higher then the max, should set to the max
     self.analytics.maxWait = kMaxWaitSeconds + 1;
-    STAssertEquals(self.analytics.maxWait, (NSInteger) kMaxWaitSeconds, @"maxWait is able to be set above the max value");
+    STAssertEquals(self.analytics.maxWait, kMaxWaitSeconds, @"maxWait is able to be set above the max value");
 
     // Set a value lower then then min, should set to the min
     self.analytics.maxWait = kMinWaitSeconds - 1;
-    STAssertEquals(self.analytics.maxWait, (NSInteger) kMinWaitSeconds, @"maxWait is able to be set below the min value");
+    STAssertEquals(self.analytics.maxWait, kMinWaitSeconds, @"maxWait is able to be set below the min value");
 
     // Set a value between
     self.analytics.maxWait = kMinWaitSeconds + 1;
-    STAssertEquals(self.analytics.maxWait, (NSInteger) kMinWaitSeconds + 1, @"maxWait is unable to be set to a valid value");
+    STAssertEquals(self.analytics.maxWait, kMinWaitSeconds + 1, @"maxWait is unable to be set to a valid value");
 }
 
 - (void)testSetMinBatchInterval {
     // Set a value higher then the max, should set to the max
     self.analytics.minBatchInterval = kMaxBatchIntervalSeconds + 1;
-    STAssertEquals(self.analytics.minBatchInterval, (NSInteger) kMaxBatchIntervalSeconds, @"minBatchInterval is able to be set above the max value");
+    STAssertEquals(self.analytics.minBatchInterval, kMaxBatchIntervalSeconds, @"minBatchInterval is able to be set above the max value");
 
     // Set a value lower then then min, should set to the min
     self.analytics.minBatchInterval = kMinBatchIntervalSeconds - 1;
-    STAssertEquals(self.analytics.minBatchInterval, (NSInteger) kMinBatchIntervalSeconds, @"minBatchInterval is able to be set below the min value");
+    STAssertEquals(self.analytics.minBatchInterval, kMinBatchIntervalSeconds, @"minBatchInterval is able to be set below the min value");
 
     // Set a value between
     self.analytics.minBatchInterval = kMinBatchIntervalSeconds + 1;
-    STAssertEquals(self.analytics.minBatchInterval, (NSInteger) kMinBatchIntervalSeconds + 1, @"minBatchInterval is unable to be set to a valid value");
+    STAssertEquals(self.analytics.minBatchInterval, kMinBatchIntervalSeconds + 1, @"minBatchInterval is unable to be set to a valid value");
 }
 
 - (void)testIsEventValid {
